requirements.txt
click==8.1.3
Flask==2.1.2
Flask-Login==0.6.0
Flask-WTF==1.0.1
importlib-metadata==4.11.3
is-safe-url==1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pyaml==21.10.1
PyYAML==6.0
typing_extensions==4.2.0
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
